# Performance: change user interface to reduce allocs