--- a/go.mod
+++ b/go.mod
@@ -3,9 +3,9 @@ module github.com/criteo/haproxy-spoe-go
 go 1.15
 
 require (
+	github.com/libp2p/go-buffer-pool v0.0.2
 	github.com/pkg/errors v0.9.1
 	github.com/sirupsen/logrus v1.6.0
-	github.com/stretchr/objx v0.1.1 // indirect
 	github.com/stretchr/testify v1.4.0
 	golang.org/x/sys v0.0.0-20200905004654-be1d3432aa8f // indirect
 )