--- a/spoe_test.go
+++ b/spoe_test.go
@@ -4,11 +4,13 @@ import (
 	"net"
 	"testing"
 
+	pool "github.com/libp2p/go-buffer-pool"
+	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/require"
 )
 
 func TestSPOE(t *testing.T) {
-	spoa := New(func(args []Message) ([]Action, error) {
+	spoa := New(func(msgs *MessageIterator) ([]Action, error) {
 		return nil, nil
 	})
 
@@ -29,7 +31,8 @@ func TestSPOE(t *testing.T) {
 	helloReq := helloFrame(t)
 	require.NoError(t, cod.encodeFrame(helloReq))
 
-	helloRes, ok, err := cod.decodeFrame(make([]byte, maxFrameSize))
+	helloRes := frame{}
+	ok, err := cod.decodeFrame(&helloRes)
 	require.True(t, ok)
 	require.NoError(t, err)
 
@@ -39,9 +42,57 @@ func TestSPOE(t *testing.T) {
 	notifyReq := notifyFrame(t)
 	require.NoError(t, cod.encodeFrame(notifyReq))
 
-	notifyRes, ok, err := cod.decodeFrame(make([]byte, maxFrameSize))
+	notifyRes := frame{}
+	ok, err = cod.decodeFrame(&notifyRes)
 	require.True(t, ok)
 	require.NoError(t, err)
 
 	require.Equal(t, notifyReq.streamID, notifyRes.streamID)
+}
+
+func BenchmarkSPOE(b *testing.B) {
+	log.SetLevel(log.FatalLevel)
+	spoa := New(func(msgs *MessageIterator) ([]Action, error) {
+		for msgs.Next() {
+		}
+		return nil, nil
+	})
+
+	lis, err := net.Listen("tcp", "127.0.0.1:0")
+	require.NoError(b, err)
+
+	agentError := make(chan error)
+	go func() {
+		agentError <- spoa.Serve(lis)
+	}()
+
+	client, err := net.Dial("tcp", lis.Addr().String())
+	require.NoError(b, err)
+
+	cod := newCodec(client, defaultConfig)
+
+	// hello
+	helloReq := helloFrame(b)
+	require.NoError(b, cod.encodeFrame(helloReq))
+
+	notifyReq := notifyFrame(b)
+	res := frame{}
+
+	ok, err := cod.decodeFrame(&res)
+	require.True(b, ok)
+	require.NoError(b, err)
+
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		err := cod.encodeFrame(notifyReq)
+		if err != nil {
+			b.Fatal(err)
+		}
+		_, err = cod.decodeFrame(&res)
+		if err != nil {
+			b.Error(err)
+		}
+		pool.Put(res.originalData)
+	}
 }